add quention before delete person

+++ frontend/src/pages/person/index.js
@@ -1,8 +1,15 @@
 import React, { Component } from "react";
 import api from "../../services/api";
 import { Link } from "react-router-dom";
+import { confirmAlert } from 'react-confirm-alert';
+import { ToastContainer } from "react-toastify";
+
 
 import "./styles.css";
+import 'react-confirm-alert/src/react-confirm-alert.css';
+
+import Notifications from "../../components/Notifications";
+
 
 export default class Person extends Component {
   state = {
@@ -19,7 +26,6 @@ export default class Person extends Component {
 
   async componentDidMount() {
     this.loadPerson();
-    this.deletePerson()
   }
 
   loadPerson = async () => { 
@@ -30,15 +36,37 @@ export default class Person extends Component {
     this.setState({ person: response.data });
   };
 
-  deletePerson = async () => { 
-    const { id } = this.state.person;
-
-    const response = await api.delete(`people/delete/${id}`);
-
-    console.log(response);
+  confirmDelete = async () => { 
+    console.log('dentro confirm');
     
+    confirmAlert({
+      title: 'Atenção',
+      message: 'Tem certeza que deseja excluir esse cadastro?',
+      buttons: [
+        {
+          label: 'Sim',
+          onClick: () => {
+            const { id } = this.state.person;
+  
+            const response = await api.delete(`people/delete/${id}`);
+            console.log(response);
+            Notifications("success", "Cadastro excluido com sucesso!");
+          }
+          
+        },
+        {
+          label: 'Não',
+          onClick: () => {
+          }
+        }
+      ]
+    });    
   };
 
+  deletePerson = async () => {
+    
+  }
+  
   render() {
     const { person } = this.state;
     var status;
@@ -80,11 +108,12 @@ export default class Person extends Component {
           <Link to="/" className="linkReturn">
             <button className="btnReturn">Voltar</button>
           </Link>
-          <button className="btnDelete" onClick={this.deletePerson}>Excluir</button>
+          <button className="btnDelete" onClick={this.confirmDelete}>Excluir</button>
           <Link to="/update/person" className="linkEdit">
             <button>Editar</button>
           </Link>
         </div>
+        <ToastContainer />
       </div>
     );
   }
